# src/contexts/marketplace/booking/presentation/booking-chat.controller.ts
import {
  Body,
  Controller,
  Get,
  HttpCode,
  HttpStatus,
  Param,
  Post,
  Put,
  Query,
  UseGuards,
} from '@nestjs/common';
import { ApiBearerAuth, ApiOperation, ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger';
import { JwtAuthGuard } from '@common/guards/jwt-auth.guard';
import { CurrentUser } from '@common/decorators/current-user.decorator';
import { AuthUser } from '@shared/types/auth-user.type';
import { BookingChatService } from '../application/services/booking-chat.service';
import {
  BookingMessageResponseDto,
  SendBookingMessageDto,
} from '../application/dto/booking-message.dto';

@ApiTags('Service Bookings')
@ApiBearerAuth()
@UseGuards(JwtAuthGuard)
@Controller('bookings/:id/messages')
export class BookingChatController {
  constructor(private readonly bookingChatService: BookingChatService) {}

  @Post()
  @ApiOperation({ summary: 'Send a booking workspace message' })
  @ApiResponse({ status: 201, type: BookingMessageResponseDto })
  async send(
    @Param('id') bookingId: string,
    @CurrentUser() user: AuthUser,
    @Body() dto: SendBookingMessageDto,
  ): Promise<BookingMessageResponseDto> {
    return this.bookingChatService.send(bookingId, user.id, dto);
  }

  @Get()
  @ApiOperation({ summary: 'List booking workspace messages' })
  @ApiQuery({ name: 'page', required: false })
  @ApiQuery({ name: 'limit', required: false })
  @ApiResponse({ status: 200, type: [BookingMessageResponseDto] })
  async list(
    @Param('id') bookingId: string,
    @CurrentUser() user: AuthUser,
    @Query('page') page?: number,
    @Query('limit') limit?: number,
  ): Promise<BookingMessageResponseDto[]> {
    return this.bookingChatService.list(bookingId, user.id, page, limit);
  }

  @Put(':msgId/read')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ summary: 'Mark a booking workspace message as read' })
  @ApiResponse({ status: 204 })
  async markAsRead(
    @Param('id') bookingId: string,
    @Param('msgId') messageId: string,
    @CurrentUser() user: AuthUser,
  ): Promise<void> {
    return this.bookingChatService.markAsRead(bookingId, messageId, user.id);
  }
}