# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/profiles/me": {
      "get": {
        "operationId": "ProfileController_getMyProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "User Profiles"
        ]
      },
      "put": {
        "operationId": "ProfileController_updateMyProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "User Profiles"
        ]
      }
    },
    "/api/profiles/{userId}": {
      "get": {
        "operationId": "ProfileController_getPublicProfile",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get public profile by user ID",
        "tags": [
          "User Profiles"
        ]
      }
    },
...
    "/api/provider-profiles/me": {
      "get": {
        "operationId": "ProviderProfileController_getMyProviderProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user provider profile",
        "tags": [
          "Provider Profiles"
        ]
      },
      "put": {
        "operationId": "ProviderProfileController_updateMyProviderProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user provider profile",
        "tags": [
          "Provider Profiles"
        ]
      }
    },
    "/api/provider-profiles/me/portfolio": {
      "post": {
        "operationId": "ProviderProfileController_addPortfolioItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an item to portfolio",
        "tags": [
          "Provider Profiles"
        ]
      }
    },
    "/api/provider-profiles/me/portfolio/{index}": {
      "delete": {
        "operationId": "ProviderProfileController_removePortfolioItem",
        "parameters": [
          {
            "name": "index",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove an item from portfolio",
        "tags": [
          "Provider Profiles"
        ]
      }
    },
    "/api/provider-profiles/me/certifications": {
      "post": {
        "operationId": "ProviderProfileController_addCertification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CertificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a certification",
        "tags": [
          "Provider Profiles"
        ]
      }
    },
    "/api/provider-profiles/me/availability": {
      "put": {
        "operationId": "ProviderProfileController_updateAvailability",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update availability status",
        "tags": [
          "Provider Profiles"
        ]
      }
    },
    "/api/provider-profiles/search": {
      "get": {
        "operationId": "ProviderProfileController_searchProviders",
        "parameters": [
          {
            "name": "skills",
            "required": false,
            "in": "query",
            "description": "Comma-separated list of skills",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minRating",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxHourlyRate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAvailable",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Search for service providers",
        "tags": [
          "Provider Profiles"
        ]
      }
    },
    "/api/provider-profiles/{userId}": {
      "get": {
        "operationId": "ProviderProfileController_getPublicProviderProfile",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderProfileResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get public provider profile",
        "tags": [
          "Provider Profiles"
        ]
      }
    },
...
    "/api/search": {
      "get": {
        "operationId": "DiscoveryController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "services",
                "profiles",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Global search across services and profiles",
        "tags": [
          "Discovery"
        ]
      }
    },
...
    "/api/bookings/{id}/files/{fileId}": {
      "get": {
        "operationId": "BookingFileController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingFileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get booking workspace file details",
        "tags": [
          "Service Bookings"
        ]
      },
      "delete": {
        "operationId": "BookingFileController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a booking workspace file",
        "tags": [
          "Service Bookings"
        ]
      }
    },
    "/api/bookings/{id}/messages": {
      "post": {
        "operationId": "BookingChatController_send",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendBookingMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingMessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a booking workspace message",
        "tags": [
          "Service Bookings"
        ]
      },
      "get": {
        "operationId": "BookingChatController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingMessageResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List booking workspace messages",
        "tags": [
          "Service Bookings"
        ]
      }
    },
    "/api/bookings/{id}/messages/{msgId}/read": {
      "put": {
        "operationId": "BookingChatController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "msgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a booking workspace message as read",
        "tags": [
          "Service Bookings"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gripper API",
    "description": "The comprehensive API for the Gripper platform backend.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://api.gripper.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "birthdate": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other",
              "prefer_not_to_say"
            ]
          },
          "location": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "socialLinks": {
            "type": "object"
          },
          "verificationStatus": {
            "type": "string"
          },
          "isVerified": {
            "type": "boolean"
          },
          "strikeCount": {
            "type": "number"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "followersCount": {
            "type": "number"
          },
          "followingCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "latestVerification": {
            "$ref": "#/components/schemas/VerificationResponseDto"
          },
          "verificationHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VerificationResponseDto"
            }
          }
        },
        "required": [
          "id",
          "userId",
          "verificationStatus",
          "isVerified",
          "strikeCount",
          "isFeatured",
          "followersCount",
          "followingCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "displayName": {
            "type": "string",
            "example": "JohnD"
          },
          "bio": {
            "type": "string",
            "example": "I am a software engineer"
          },
          "avatarUrl": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          },
          "coverImageUrl": {
            "type": "string",
            "example": "https://example.com/cover.jpg"
          },
          "birthdate": {
            "type": "string",
            "example": "1990-01-01"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other",
              "prefer_not_to_say"
            ]
          },
          "location": {
            "type": "string",
            "example": "New York"
          },
          "country": {
            "type": "string",
            "example": "USA"
          },
          "city": {
            "type": "string",
            "example": "New York City"
          },
          "website": {
            "type": "string",
            "example": "https://example.com"
          },
          "socialLinks": {
            "type": "object"
          }
        }
      },
...
      "CertificationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Full Stack Web Development"
          },
          "issuer": {
            "type": "string",
            "example": "Udemy"
          },
          "year": {
            "type": "number",
            "example": 2023
          },
          "url": {
            "type": "string",
            "example": "https://cert.com/123"
          }
        },
        "required": [
          "name",
          "issuer",
          "year"
        ]
      },
      "PortfolioItemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "E-commerce Website"
          },
          "description": {
            "type": "string",
            "example": "A complete online store built with NestJS"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/project.jpg"
          },
          "projectUrl": {
            "type": "string",
            "example": "https://github.com/project"
          }
        },
        "required": [
          "title",
          "description",
          "imageUrl"
        ]
      },
      "ProviderProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "businessName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "certifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CertificationDto"
            }
          },
          "yearsOfExperience": {
            "type": "number"
          },
          "portfolio": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioItemDto"
            }
          },
          "hourlyRate": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "responseTimeHours": {
            "type": "number"
          },
          "completionRate": {
            "type": "number"
          },
          "totalEarnings": {
            "type": "number"
          },
          "totalJobs": {
            "type": "number"
          },
          "averageRating": {
            "type": "number"
          },
          "totalReviews": {
            "type": "number"
          },
          "isAvailable": {
            "type": "boolean"
          },
          "availabilityHours": {
            "type": "object"
          },
          "lastActiveAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "latestVerification": {
            "$ref": "#/components/schemas/VerificationResponseDto"
          }
        },
        "required": [
          "id",
          "userId",
          "completionRate",
          "totalEarnings",
          "totalJobs",
          "averageRating",
          "totalReviews",
          "isAvailable",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProviderProfileDto": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string",
            "example": "Doe Tech Solutions"
          },
          "description": {
            "type": "string",
            "example": "Providing high-quality software development services."
          },
          "skills": {
            "example": [
              "JavaScript",
              "NestJS",
              "React"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "certifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CertificationDto"
            }
          },
          "yearsOfExperience": {
            "type": "number",
            "example": 5
          },
          "portfolio": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioItemDto"
            }
          },
          "hourlyRate": {
            "type": "number",
            "example": 50
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "responseTimeHours": {
            "type": "number",
            "example": 24
          },
          "isAvailable": {
            "type": "boolean",
            "example": true
          },
          "availabilityHours": {
            "type": "object",
            "example": {
              "Monday": {
                "start": "09:00",
                "end": "17:00"
              }
            }
          }
        }
      },
      "UpdateAvailabilityDto": {
        "type": "object",
        "properties": {
          "isAvailable": {
            "type": "boolean"
          }
        },
        "required": [
          "isAvailable"
        ]
      },
...
      "BookingFileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bookingId": {
            "type": "string"
          },
          "uploadedBy": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "originalName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "url": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "bookingId",
          "uploadedBy",
          "filename",
          "originalName",
          "mimeType",
          "size",
          "url",
          "fileType",
          "createdAt"
        ]
      },
      "SendBookingMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "Please check the latest upload."
          },
          "messageType": {
            "type": "string",
            "example": "text"
          },
          "attachments": {
            "example": [
              "https://cdn.example.com/file.pdf"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        },
        "required": [
          "content"
        ]
      },
      "BookingMessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bookingId": {
            "type": "string"
          },
          "senderId": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "messageType": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "readAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "bookingId",
          "senderId",
          "content",
          "messageType",
          "createdAt"
        ]
      },
...
    }
  }
}